requirements.txt
# API and HTTP handling
httpx==0.26.0
tenacity==8.2.3

# Database
psycopg2-binary==2.9.9
SQLAlchemy>=1.4.28,<2.0
alembic==1.13.1

# Data validation and processing
pydantic==2.6.1
pandas==2.2.0

# Secrets management
hvac==2.1.0  # HashiCorp Vault client

# Orchestration
apache-airflow==2.8.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Logging and monitoring
python-json-logger==2.0.7

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0 